backend/requirements.txt
asgiref==3.8.1
Django==5.2
django-admin-filters==0.1.3
django-advanced-filters==2.0.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
PyJWT==2.9.0
python-dotenv==1.1.0
simplejson==3.20.1
sqlparse==0.5.3
tzdata==2025.2
